Clones now have an end to their timeline. Ignored energy fields in level 5 for objects in motion fixed colliders and buttons in main menu

+++ ChronoLoop/ChronoLoop/Actions/CCCloneIndicator.h
@@ -13,7 +13,7 @@ namespace Epoch
 
 		CCCloneIndicator(unsigned int _id)
 		{
-			mSnap = TimeManager::Instance()->GetCurrentSnapFrame();
+			mSnap = (float)TimeManager::Instance()->GetCurrentSnapFrame();
 			mId = _id;
 		}
